discountable and refundable added

+++ app/src/main/java/com/example/louisereid/stevesmusicexchange/Behaviours/Sellable.java
@@ -6,5 +6,5 @@ package com.example.louisereid.stevesmusicexchange.Behaviours;
 
 public interface Sellable {
 
-    int calcMarkUp();
+    double calcMarkUp();
 }
